Support turning off ncurses support in cursive
The cursive crate turns on ncurses support by default. When we depend on it without setting `default-features = false`, we force this support to always be turned on without giving downstream users the ability to turn it off.

This commit disables the default features in the dependency, but **our** default features still turn on ncurses support.